requirements.txt
streamlit
pandas
matplotlib
gspread
google-auth